+++ src/main/java/de/autumnal/discordmusicbotnetwork/message/enums/Commands.java
@@ -1,0 +1,6 @@
+package de.autumnal.discordmusicbotnetwork.message.enums;
+
+public enum Commands {
+    FORWARD,
+    JUMP
+}