+++ src/main/java/de/autumnal/discordmusicbotnetwork/message/enums/Language.java
@@ -1,0 +1,15 @@
+package de.autumnal.discordmusicbotnetwork.message.enums;
+
+public enum Language {
+    ENGLISCH,
+    GERMAN;
+
+    public static Language getLanguageFromString(String language){
+        language = language.toUpperCase();
+        try {
+            return Language.valueOf(language);
+        }catch (Exception e){
+            return Language.ENGLISCH;
+        }
+    }
+}