+++ src/main/java/de/autumnal/discordmusicbotnetwork/music/enums/AudioPlayerState.java
@@ -1,0 +1,7 @@
+package de.autumnal.discordmusicbotnetwork.music.enums;
+
+public enum AudioPlayerState {
+    PLAYING,
+    PAUSED,
+    STOPPED
+}