+++ src/main/java/de/autumnal/discordmusicbotnetwork/music/enums/Playmode.java
@@ -1,0 +1,7 @@
+package de.autumnal.discordmusicbotnetwork.music.enums;
+
+public enum Playmode {
+    NORMAL,
+    LOOPONE,
+    LOOPALL;
+}