+++ src/main/java/message/enums/Commands.java
@@ -1,6 +1,0 @@
-package message.enums;
-
-public enum Commands {
-    FORWARD,
-    JUMP
-}