+++ src/main/java/music/enums/AudioPlayerState.java
@@ -1,7 +1,0 @@
-package music.enums;
-
-public enum AudioPlayerState {
-    PLAYING,
-    PAUSED,
-    STOPPED
-}